feat: 改进ignore_bubble功能，阈值改为0-1范围
改进：
- 阈值从1-50改为0-1（0=禁用，0.01-0.3=宽松，0.3-0.7=中等，0.7-1.0=严格）
- 保留原有简单边缘检测方法
- 添加高级气泡边界检测方法（基于旧版本逻辑）
- 更新配置模型：ignore_bubble从int改为float
- 更新文档说明新的阈值范围和效果
- 更新Qt UI和Server UI的输入控件

+++ desktop_qt_ui/core/config_models.py
@@ -20,7 +20,7 @@ class OcrSettings(BaseModel):
     use_hybrid_ocr: bool = True
     secondary_ocr: str = "mocr"
     min_text_length: int = 0
-    ignore_bubble: int = 0
+    ignore_bubble: float = 0.0
     prob: float = 0.1
     merge_gamma: float = 0.8
     merge_sigma: float = 2.5
